{{cookiecutter.project_slug}}/requirements.txt
django
pytz
djangorestframework
django-filter
{% if cookiecutter.use_sentry == 'y' -%}
raven
{%- endif %}
